Py2 Compatibilty, folder structure, encoding
* ensured files are utf-8 encoded for Py2
* Fixed some Python 2 compatibility issues with the
TimedCache
* Updates to README
