+++ docs/pydecor._memoization.rst
@@ -1,0 +1,8 @@
+pydecor\.\_memoization module
+=============================
+
+.. automodule:: pydecor._memoization
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
