+++ docs/pydecor.decorators._utility.rst
@@ -1,0 +1,8 @@
+pydecor\.decorators\.\_utility module
+=====================================
+
+.. automodule:: pydecor.decorators._utility
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
