+++ docs/pydecor.decorators.generic.rst
@@ -1,0 +1,8 @@
+pydecor\.decorators\.generic module
+===================================
+
+.. automodule:: pydecor.decorators.generic
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
