+++ docs/pydecor.decorators.ready_to_wear.rst
@@ -1,0 +1,8 @@
+pydecor\.decorators\.ready\_to\_wear module
+===========================================
+
+.. automodule:: pydecor.decorators.ready_to_wear
+    :members:
+    :private-members:
+    :undoc-members:
+    :show-inheritance:
