+++ docs/pydecor.decorators.rst
@@ -1,5 +1,17 @@
-pydecor\.decorators module
-==========================
+pydecor\.decorators package
+===========================
+
+Submodules
+----------
+
+.. toctree::
+
+   pydecor.decorators._utility
+   pydecor.decorators.generic
+   pydecor.decorators.ready_to_wear
+
+Module contents
+---------------
 
 .. automodule:: pydecor.decorators
     :members:
