+++ pydecor/_version.py
@@ -1,3 +1,4 @@
+# -*- coding: UTF-8 -*-
 """
 version.py module
 
@@ -7,5 +8,5 @@ and also set as the __version__ attribute for the package.
 
 from __future__ import absolute_import, unicode_literals
 
-__version_info__ = (1, 0, 0)
+__version_info__ = (1, 1, 0)
 __version__ = '.'.join([str(ver) for ver in __version_info__])
