+++ pydecor/caches.py
@@ -1,14 +1,34 @@
+# -*- coding: UTF-8 -*-
 """
 Caches for memoization
 """
 
+from __future__ import absolute_import, unicode_literals
+
+
+__all__ = (
+    'LRUCache',
+    'FIFOCache',
+    'TimedCache',
+)
+
 
 from collections import OrderedDict
-from datetime import datetime
+from time import time
 
 
 class LRUCache(OrderedDict):
-    """Remove least-recently used items
+    """Self-pruning cache using an LRU strategy
+
+    If instantiated with a ``max_size`` other than ``0``, will
+    automatically prune the least-recently-used (LRU) key/value
+    pair when inserting an item after reaching the specified size.
+
+    An item is considered to be "used" when it is inserted or
+    accessed, at which point its position in recently used
+    queue is updated to the most recent.
+
+    Supports all standard dictionary methods.
 
     :param int max_size: maximum number of entries to save
         before pruning
@@ -33,7 +53,14 @@ class LRUCache(OrderedDict):
 
 
 class FIFOCache(OrderedDict):
-    """Removes the first input item when full
+    """Self-pruning cache using a FIFO strategy
+
+    If instantiated with a ``max_size`` other than ``0``, will
+    automatically prune the least-recently-inserted key/value
+    pair when inserting an item after reaching the specified
+    size.
+
+    Supports all standard dictionary methods.
 
     :param int max_size: maximum number of entries to save
         before pruning
@@ -50,12 +77,14 @@ class FIFOCache(OrderedDict):
 
 
 class TimedCache(dict):
-    """Remove items older than the specified time
+    """Self-pruning cache whose entries can be set to expire
 
-    When trying to access an entry, either via ``in`` or
-    ``__getitem__``, if it is older than the specified
-    ``max_age``, in seconds, it will be deleted, and the
-    data structure will act as though it does not exist.
+    If instantiated with a ``max_age`` other than ``0``, will
+    consider entries older than the specified age to be invalid,
+    removing them from the cache upon an attempt to access them
+    and returning as though they do not exist.
+
+    Supports all standard dictionary methods.
 
     :param int max_age: age in seconds beyond which entries
         should be considered invalid. The default is 0, which
@@ -68,7 +97,7 @@ class TimedCache(dict):
 
     def __getitem__(self, key):
         value, last_time = dict.__getitem__(self, key)
-        now = datetime.utcnow().timestamp()
+        now = time()
         if self._max_age and now - last_time > self._max_age:
             del self[key]
             raise KeyError(key)
@@ -76,7 +105,7 @@ class TimedCache(dict):
             return value
 
     def __setitem__(self, key, value):
-        now = datetime.utcnow().timestamp()
+        now = time()
         dict.__setitem__(self, key, (value, now))
 
     def __contains__(self, key):
