+++ pydecor/decorators/__init__.py
@@ -1,0 +1,20 @@
+# -*- coding: UTF-8 -*-
+"""
+Public interface for decorators sub-package
+"""
+
+__all__ = (
+    'after',
+    'before',
+    'construct_decorator',
+    'decorate',
+    'instead',
+    'Decorated',
+    'DecoratorType',
+    'intercept',
+    'log_call',
+    'memoize',
+)
+
+from .generic import *
+from .ready_to_wear import *
