+++ requirements-dev.txt
@@ -5,3 +5,4 @@ pytest
 pytest-cov
 sphinx
 sphinx-autodoc-typehints
+tox
